Player events and Forge events run in their own separate buses without redundant registering.
